Refactor ExitManager to improve filtering
Reworked the ExitManager to enhance the logic filtering exit points based on the map and player.

TODO: Fix name mapping for extracts, as some exfils are currently not shown due to mismatches  between in-game names and data source. Further investigation and mapping adjustments are needed to ensure all relevant exits are displayed.